Update all consumers of PhaseFunction::p to use Spectrum
PBRT assumed that the output from a Phase Function's phase(p) is a Float, but in the case of the new tabulated version the output is a Spectrum. Surprisingly, the output of `p' is  converted to a SampledSpectrum either explicitly or implicitly. This change forces all calls to be explicit by returning a Spectrum.