Replace Medium in Ray with MediaTracker
Because the wavefront integrator (through the SOA code) will not allow for a dynamically allocated Mediums on a Ray, we create a new class, MediaTracker, to be used in its stead. The number of potentially nested mediums/volumes is set at compile time.

Right now, this class acts no differently than the ray.medium version as we are only using the first element of the array for Mediums.