Fix overwriting valid light intersection point in ray export
In SampleLD, when an intersection ray misses, it means there was nothing obstructing the light. In that case, we add the light point as the final in the path, and exit. The code in the main loop assumed that *all* current segments ended with the bogus point marker, and overrode the valid point. Now check that the final point in the segment is bogus before overwriting.