Enable removing of any volume, not just the last one
Under some circumstances, the ray may exit a volume that was not the last pushed, e.g. overlapping and not nested. This was showing up as a warning about exceeding the maximum number of volumes. This commit allows for the removal of the volume anywhere in the pushed list, and then compacts the list.