Test for edge cases where we enter but never exit a volume
This can occur when the ray intersects the corner of a volume. The ray is marked as "entering", but due to numerical precision, or hitting the exact edge, never exits the volume. Sample_TMaj catches this case immediately, so the ray never contributes to the computation of TMaj, but it can interfere with the computation of how many volumes are on the stack. So, when a new volume is pushed and the stack exceeds the predetermined capacity, we check the intersection point against all of the active volumes and remove any that are bogus.