Fix bug in multi-volume traversal
Under certain conditions, the PBRT MajorantGrid would appear while rendering. The issue comes from how the original code was striding through the overlapping grids resulting in skipping cells and backtracking. The image would display outlines of the grid and errors in volume processing such as a back volume not be composited with a front volume.

The fix was to use the MajorantGrid with the smallest step size for striding, but use the stochastically sampled volume for media lookup values.

A number of other fixes were implemented, including:
- Wrong indexing for updating location in non-selected volumes
- A better method for updating the location in non-selected volumes